refactor: covert index, main to TypeScript
