--- src/helper/index.js
@@ -1,3 +1,0 @@
-export { default as defaultProps } from './defaultProps.js'
-export { redColorScaleGenerator } from './redColorScaleGenerator'
-export * from './utilities.js'
