+++ src/helper/index.ts
@@ -1,0 +1,3 @@
+export { default as defaultProps } from './defaultProps'
+export { redColorScaleGenerator } from './redColorScaleGenerator'
+export * from './utilities'
