--- src/index.js
@@ -1,11 +1,0 @@
-import { VuePivottable, VuePivottableUi } from './components'
-import TableRenderer from './components/pivottable/renderer'
-import * as PivotUtilities from './helper'
-export * from './composables'
-
-const Renderer = {
-  ...TableRenderer
-}
-
-export { VuePivottable, VuePivottableUi, PivotUtilities, Renderer }
-export default { VuePivottable, VuePivottableUi, PivotUtilities, Renderer }
