+++ src/index.ts
@@ -1,8 +1,8 @@
 import { VuePivottable, VuePivottableUi } from './components'
 import TableRenderer from './components/pivottable/renderer'
 import * as PivotUtilities from './helper'
-export * from './composables/index.js'
 import type { Component } from 'vue'
+export * from './composables'
 
 const Renderer: Record<string, Component> = {
   ...TableRenderer
